backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
motor==3.3.1
certifi>=2024.2.2
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
python-multipart>=0.0.9
httpx>=0.27.0
pillow>=10.2.0
aiohttp>=3.9.0

# AI - Lightweight (using Hugging Face API)
google-generativeai>=0.5.0
